refactor: クラス名とファイル名が一致するよう修正
