+++ 04.class/memoApp/memoDeletion.js
