+++ 04.class/memoApp/memoDisplay.js
